#4237 changelly 2

+++ mbw/src/main/java/com/mycelium/wallet/external/changelly/model/FixRate.kt
@@ -1,5 +1,7 @@
 package com.mycelium.wallet.external.changelly.model
 
+import java.math.BigDecimal
+
 
 //"id": "f4dd43106d63b65b88955a0b362645ce960987c7ffb7a8480dd32e799431177f",
 //"result": "0.02556948",
@@ -11,10 +13,10 @@ package com.mycelium.wallet.external.changelly.model
 //"minTo": "0.00379488"
 
 class FixRate(val id: String,
-              val result: Double,
+              val result: BigDecimal,
               val from: String,
               val to: String,
-              val maxFrom: Double,
-              val maxTo: Double,
-              val minFrom: Double,
-              val minTo: Double)
+              val maxFrom: BigDecimal,
+              val maxTo: BigDecimal,
+              val minFrom: BigDecimal,
+              val minTo: BigDecimal)
